Added EsacChecksumManagement class to deal with checksums stored locally at ESAC
