+++ caom2-artifact-sync/src/main/java/ca/nrc/cadc/caom2/artifactsync/ESACArtifactStore.java
@@ -84,22 +84,13 @@ import java.nio.channels.FileChannel.MapMode;
 import java.nio.file.Paths;
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.sql.Statement;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import org.apache.log4j.Level;
 import org.apache.log4j.Logger;
 
-import ca.nrc.cadc.caom2.artifactsync.esacdb.ConfigProperties;
-import ca.nrc.cadc.caom2.artifactsync.esacdb.JdbcSingleton;
-import ca.nrc.cadc.caom2.artifactsync.persistance.EsacChecksumPersistance;
+import ca.nrc.cadc.caom2.artifactsync.persistance.EsacChecksumManagement;
 import ca.nrc.cadc.net.TransientException;
 import ca.nrc.cadc.util.StringUtil;
 
@@ -115,35 +106,18 @@ public class ESACArtifactStore implements ArtifactStore
 {
     private static final Logger log = Logger.getLogger(ESACArtifactStore.class);
 
-    private PreparedStatement preparedStatement = null;
-
-    protected static final String SCHEMA = "caom2.artifactsync.checksumTable.schema";
-    protected static final String TABLENAME = "caom2.artifactsync.checksumTable.table";
-    protected static final String COLUMN_ARTIFACT = "caom2.artifactsync.checksumTable.columnArtifact";
-    protected static final String COLUMN_CHECKSUM = "caom2.artifactsync.checksumTable.columnChecksum";
-    private static String checksumSchema = null;
-    private static String checksumTable = null;
-    private static String checksumArtifactColumnName = null;
-    private static String checksumChecksumColumnName = null;
-
     String dataURL;
 
     public ESACArtifactStore()
             throws IllegalArgumentException, SecurityException, IllegalAccessException, InvocationTargetException, NoSuchMethodException, SQLException, PropertyVetoException
     {
-        ESACArtifactStore.checksumSchema = ConfigProperties.getInstance().getProperty(SCHEMA);
-        ESACArtifactStore.checksumTable = ConfigProperties.getInstance().getProperty(TABLENAME);
-        ESACArtifactStore.checksumArtifactColumnName = ConfigProperties.getInstance().getProperty(COLUMN_ARTIFACT);
-        ESACArtifactStore.checksumChecksumColumnName = ConfigProperties.getInstance().getProperty(COLUMN_CHECKSUM);
-
-        preparedStatement = createPreparedStatement(JdbcSingleton.getInstance().getConnection(), checksumSchema, checksumTable);
     }
 
     @Override
     public boolean contains(URI artifactURI, URI checksum) throws TransientException
     {
         init();
-        boolean result = select(artifactURI, checksum);
+        boolean result = EsacChecksumManagement.getInstance().select(artifactURI, checksum);
         return result;
     }
 
@@ -155,11 +129,11 @@ public class ESACArtifactStore implements ArtifactStore
         {
             if (!contains(artifactURI))
             {
-                insert(artifactURI, checksum);
+                EsacChecksumManagement.getInstance().insert(artifactURI, checksum);
             }
             else
             {
-                update(artifactURI, checksum);
+                EsacChecksumManagement.getInstance().update(artifactURI, checksum);
             }
         }
     }
@@ -171,176 +145,8 @@ public class ESACArtifactStore implements ArtifactStore
 
     private boolean contains(URI artifactURI) throws TransientException
     {
-        boolean result = select(artifactURI);
+        boolean result = EsacChecksumManagement.getInstance().select(artifactURI);
         return result;
-    }
-
-    private boolean select(URI artifactURI)
-    {
-        boolean result = false;
-        Statement stmt = null;
-        String query = "select * from " + checksumSchema + "." + checksumTable + " where " + checksumArtifactColumnName + " = '" + artifactURI.toString() + "';";
-        ResultSet rs = null;
-
-        try
-        {
-            Connection con = JdbcSingleton.getInstance().getConnection();
-            stmt = con.createStatement();
-            rs = stmt.executeQuery(query);
-            if (rs.next())
-            {
-                result = true;
-            }
-            else
-            {
-                result = false;
-            }
-
-        }
-        catch (Exception ex)
-        {
-            throw new RuntimeException("Unexpected exception: " + ex.getMessage());
-        }
-        finally
-        {
-            if (rs != null)
-            {
-                try
-                {
-                    rs.close();
-                }
-                catch (SQLException e)
-                {
-                }
-            }
-            if (stmt != null)
-            {
-                try
-                {
-                    stmt.close();
-                }
-                catch (SQLException e)
-                {
-                }
-            }
-        }
-        return result;
-
-    }
-
-    private boolean select(URI artifactURI, URI checksum)
-    {
-        boolean result = false;
-        Statement stmt = null;
-        String query = "select * from " + checksumSchema + "." + checksumTable + " where " + checksumArtifactColumnName + " = '" + artifactURI.toString() + "' and "
-                + checksumChecksumColumnName + " = '" + checksum.toString() + "'";
-        ResultSet rs = null;
-
-        try
-        {
-            Connection con = JdbcSingleton.getInstance().getConnection();
-            stmt = con.createStatement();
-            rs = stmt.executeQuery(query);
-            if (rs.next())
-            {
-                result = true;
-            }
-            else
-            {
-                result = false;
-            }
-
-        }
-        catch (Exception ex)
-        {
-            throw new RuntimeException("Unexpected exception: " + ex.getMessage());
-        }
-        finally
-        {
-            if (rs != null)
-            {
-                try
-                {
-                    rs.close();
-                }
-                catch (SQLException e)
-                {
-                }
-            }
-            if (stmt != null)
-            {
-                try
-                {
-                    stmt.close();
-                }
-                catch (SQLException e)
-                {
-                }
-            }
-        }
-        return result;
-    }
-
-    private void insert(URI artifactURI, URI checksum)
-    {
-        try
-        {
-            List<EsacChecksumPersistance> list = new ArrayList<EsacChecksumPersistance>();
-            list.add(new EsacChecksumPersistance(artifactURI, checksum));
-            databaseWriter(preparedStatement, list);
-        }
-        catch (Exception ex)
-        {
-            throw new RuntimeException("Unexpected exception: " + ex.getMessage());
-        }
-
-    }
-
-    private void update(URI artifactURI, URI checksum)
-    {
-        Statement stmt = null;
-        String update = "update " + checksumSchema + "." + checksumTable + " set " + checksumChecksumColumnName + " = '" + checksum.toString() + "' where "
-                + checksumArtifactColumnName + " = '" + artifactURI.toString() + "';";
-        ResultSet rs = null;
-
-        try
-        {
-            Connection con = JdbcSingleton.getInstance().getConnection();
-            stmt = con.createStatement();
-            int res = stmt.executeUpdate(update);
-            if (res != 1)
-            {
-                throw new RuntimeException("Unexpected exception updating artifact: " + artifactURI.toString());
-            }
-        }
-        catch (Exception ex)
-        {
-            throw new RuntimeException("Unexpected exception: " + ex.getMessage());
-        }
-        finally
-        {
-            if (rs != null)
-            {
-                try
-                {
-                    rs.close();
-                }
-                catch (SQLException e)
-                {
-                }
-            }
-            if (stmt != null)
-            {
-                try
-                {
-                    stmt.close();
-                }
-                catch (SQLException e)
-                {
-                }
-            }
-        }
-
     }
 
     private String calculateMD5Sum(String path) throws UnsupportedOperationException
@@ -404,40 +210,6 @@ public class ESACArtifactStore implements ArtifactStore
         {
             throw new IllegalStateException("BUG in forming data URL", e);
         }
-    }
-
-    private static PreparedStatement createPreparedStatement(Connection con, String schema, String tableName)
-            throws IllegalArgumentException, SecurityException, IllegalAccessException, InvocationTargetException, NoSuchMethodException, SQLException
-    {
-        PreparedStatement insertValues = null;
-
-        String table = schema + "." + tableName;
-
-        String columns = checksumArtifactColumnName + "," + checksumChecksumColumnName;
-        String values = "?,?";
-
-        String insertStatement = "insert into " + table + "(" + columns + ")" + " values (" + values + ");";
-
-        log.log(Level.INFO, "Insertion prepared statement " + insertStatement);
-
-        insertValues = con.prepareStatement(insertStatement);
-
-        return insertValues;
-    }
-
-    private static int databaseWriter(PreparedStatement insertValues, List<EsacChecksumPersistance> sources) throws Exception
-    {
-        int idx = 1;
-        for (EsacChecksumPersistance source : sources)
-        {
-            insertValues.setString(idx++, source.getArtifactURI().toString());
-            insertValues.setString(idx++, source.getChecksum().toString());
-        }
-        insertValues.addBatch();
-        insertValues.executeBatch();
-        insertValues.getConnection().commit();
-
-        return sources.size();
     }
 
     class ESACDataURI
